requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.12.0
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
